+++ rio/ChangeLog.md
@@ -3,6 +3,8 @@
 ## 0.1.3.0
 
 * Allow dynamic reloading of `logMinLevel` and `logVerboseFormat` for the `LogOptions` record
+* Add `SimpleApp` and `runRIOSimple`
+* Add `asIO`
 
 ## 0.1.2.0